pyproject.toml
[project]
name = "firecrawl-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firecrawl-py>=4.5.0",
    "python-dotenv>=1.1.1",
]
